+++ src/game.js
@@ -3,7 +3,7 @@ import TWEEN from "tween.js";
 import { ElectronRuntime, SteamRuntime } from "@/steam";
 
 import { DC } from "./core/constants";
-import { deepmergeAll } from "@/utility/deepmerge";
+import { deepmergeAll } from "@/utils/deepmerge";
 import { DEV } from "@/env";
 import { SpeedrunMilestones } from "./core/speedrun";
 import { supportedBrowsers } from "./supported-browsers";
