+++ src/utils/open-external-link.js
