+++ src/utils/type-check.js
