+++ src/utils/word-shift.js
